CPU & GPU Analysis (Cartesian)
xy座標での解析実行
GPU側の0-step capture regionは正しく計算できていた
CPUとGPUの解析結果が一致した
CPUのgrid.cppに致命的なバグがあったので修正

1-step以上の結果確認はまだ

+++ capturability/include/grid.h
@@ -66,7 +66,7 @@ private:
   int getStateIndexCartesian(int icp_x_id, int icp_y_id, int swf_x_id, int swf_y_id);
   int getStateIndexPolar(int icp_r_id, int icp_th_id, int swf_r_id, int swf_th_id);
 
-  int round(float value);
+  int round(double value);
   int max(int val1, int val2);
   int max(int val1, int val2, int val3, int val4);
